refactor: remove dead DomainService, rotor/tunnel STL generation, and associated CLI flags
DomainService generated rotor.stl and tunnel.stl but neither is referenced
by any template's snappyHexMeshDict. The disc template only needs disc.stl
(from StlConversionService). The blockMeshDict template variables define the
wind tunnel domain, not an STL file.

Removed:
- Services/DomainService.cs (entire file, ~400 LOC)
- DomainGenerationResult, CylinderDimensions, BoxDimensions, Vector3, Triangle
- CLI flags: --rotor-radius-scale, --rotor-height-scale, --mesh-resolution
- StudyDomainConfig: RotorRadiusScale, RotorHeightScale, MeshResolution
- TemplateMetadata: RequiresRotorZone property and backward-compat accessor
- requiresRotorZone from all TEMPLATE.json files
- DI registration for DomainService
- requiredStlFiles: removed tunnel.stl from disc template

Kept:
- BoundingBox class (used by CaseService for domain sizing)
- TunnelUpstream/Downstream/Radial in StudyDomainConfig (blockMeshDict vars)
- GeometryService facade (now wraps only StlConversionService + CalculateBoundingBox)

CaseService.ProcessGeometry now extracts bounding box directly via
GeometryService.CalculateBoundingBox instead of going through DomainService.

Build: 0 warnings, 0 errors. Tests: 243 passed, 0 failed.
