+++ Models/StudyConfig.cs
@@ -52,17 +52,11 @@ namespace foamscript.Models
     }
 
     /// <summary>
-    /// Domain geometry parameters controlling rotor zone and wind tunnel sizing.
+    /// Domain geometry parameters controlling wind tunnel sizing.
     /// All scales are relative to the detected reference dimension.
     /// </summary>
     public class StudyDomainConfig
     {
-        /// <summary>Rotor AMI cylinder radius as a multiple of geometry radius. Default: 1.25 (~25% clearance).</summary>
-        public double RotorRadiusScale { get; set; } = 1.25;
-
-        /// <summary>Rotor AMI cylinder height as a multiple of geometry height. Default: 1.5.</summary>
-        public double RotorHeightScale { get; set; } = 1.5;
-
         /// <summary>Wind tunnel upstream extent in reference lengths. Default: 5.0 (CFD convention).</summary>
         public double TunnelUpstream { get; set; } = 5.0;
 
@@ -71,8 +65,5 @@ namespace foamscript.Models
 
         /// <summary>Wind tunnel radial extent in reference lengths. Default: 5.0 (CFD convention).</summary>
         public double TunnelRadial { get; set; } = 5.0;
-
-        /// <summary>Number of vertices around the generated cylinder geometries. Default: 32.</summary>
-        public int MeshResolution { get; set; } = 32;
     }
 }
