+++ Templates/external_airfoil_static_steady/TEMPLATE.json
@@ -13,8 +13,7 @@
     "areaFormula": "rectangular"
   },
   "rotation": {
-    "enabled": false,
-    "requiresRotorZone": false
+    "enabled": false
   },
   "validation": null,
   "parameters": {
