+++ Templates/external_disc_rotating-ami_transient/TEMPLATE.json
@@ -13,8 +13,7 @@
     "areaFormula": "circular"
   },
   "rotation": {
-    "enabled": true,
-    "requiresRotorZone": true
+    "enabled": true
   },
   "validation": {
     "minSize": 0.18,
