+++ Templates/external_disc_rotatingwall_steady/TEMPLATE.json
@@ -5,7 +5,7 @@
   "geometry": {
     "type": "disc",
     "stlName": "disc.stl",
-    "requiredStlFiles": ["disc.stl", "tunnel.stl"],
+    "requiredStlFiles": ["disc.stl"],
     "surfaceOrient": { "outsidePoint": [0, 0, -1] }
   },
   "reference": {
@@ -13,8 +13,7 @@
     "areaFormula": "circular"
   },
   "rotation": {
-    "enabled": true,
-    "requiresRotorZone": true
+    "enabled": true
   },
   "validation": {
     "minSize": 0.18,
